fix(audio): retain capture callbacks and verify finalized bytes [D-AW-RECORD-01 r1 candidate]

+++ Packages/UI/Tests/DWorkbenchTests/ProjectAudioSessionTests.swift
@@ -111,7 +111,7 @@ private final class SessionAudioFactory: AudioTransportDeviceFactory {
             AudioFileWritePackets(file, false, UInt32($0.count), nil, 0, &packets, $0.baseAddress!)
         }
         let closeStatus = AudioFileClose(file)
-        guard status == noErr, packets == samples.count, closeStatus == noErr,
+        guard status == noErr, packets == UInt32(samples.count), closeStatus == noErr,
               capture.synchronize() == nil else {
             throw AudioMediaError.io("无法写入合成录音 fixture")
         }
